wasm: Phase A — TinyGo probe + baseline-build fixes
Adds wasm/main.go, a probe entrypoint for measuring the easyfl WASM
binary size under TinyGo. Building it surfaced three real issues that
also affect production hygiene:

- easyfl_util/util.go pulled github.com/stretchr/testify/require (and
  transitively net/http) into every consumer through the test helper
  RequireErrorWith. Moved the helper to a new easyfl_util/testutil
  sub-package; updated the four easyfl test callers.
- tuples/tuples.go and compiler.go had `len(x) > math.MaxUint32` /
  `n > math.MaxUint32` comparisons that overflow `int` on 32-bit
  targets (TinyGo wasm). Cast operands to uint64 for portable
  comparison; behavior on 64-bit Go is unchanged.

Baseline (full easyfl, unsplit): 1.5 MB wasm / 476 KB gzipped.

Proxima callers of easyfl_util.RequireErrorWith need updating on the
proxima side after the easyfl version bump.

+++ compiler.go
@@ -400,7 +400,7 @@ func parseLiteral[T any](lib *Library[T], sym string, w io.Writer) (bool, int, e
 		if err != nil {
 			return false, 0, fmt.Errorf("%v: '%s'", err, sym)
 		}
-		if n < 0 || n > math.MaxUint32 {
+		if n < 0 || uint64(n) > math.MaxUint32 {
 			return false, 0, fmt.Errorf("wrong u32 constant: '%s'", sym)
 		}
 		var b [4]byte
@@ -458,7 +458,7 @@ func parseLiteral[T any](lib *Library[T], sym string, w io.Writer) (bool, int, e
 		if err != nil {
 			return false, 0, fmt.Errorf("%v: '%s'", err, sym)
 		}
-		if n < 0 || n > math.MaxUint32 {
+		if n < 0 || uint64(n) > math.MaxUint32 {
 			return false, 0, fmt.Errorf("wrong z32 constant: '%s'", sym)
 		}
 		var b [4]byte
